Fix build process to create a valid output
