+++ articles/scp-3211/gulpfile.js
@@ -1,10 +1,6 @@
 const gulp = require("gulp")
 const del = require("del")
-
-const { makeFtml } = require("./src/build")
-const { langs } = require("./src/config")
-
-module.exports = gulp.series('clean', 'build')
+const ts = require("gulp-typescript")
 
 gulp.task('clean', async done => {
   // Scrub any previous builds from dist
@@ -12,10 +8,22 @@ gulp.task('clean', async done => {
   done()
 })
 
+gulp.task('make', () => {
+  const tsProject = ts.createProject("./tsconfig.json")
+  return tsProject
+    .src()
+    .pipe(tsProject())
+    .js.pipe(gulp.dest("./dist/"))
+})
+
 gulp.task('build', async done => {
   // For each configured language, build the article
+  const { makeFtml } = require("./dist/build")
+  const { langs } = require("./dist/config")
   for (const lang in langs) {
     await makeFtml(lang)
   }
   done()
 })
+
+gulp.task('default', gulp.series('clean', 'make', 'build'))
