+++ articles/scp-3211/src/en/page.ejs.ftml
@@ -34,7 +34,7 @@ iframe {
 [[/module]]
 
 [[html]]
-<%= html %>
+<%- html %>
 [[/html]]
 
 [[footnoteblock]]
