+++ articles/scp-3211/tsconfig.json
@@ -4,6 +4,7 @@
     "outDir": "./dist/",
     "strict": true,
     "module": "commonjs",
+    "target": "es5",
     "moduleResolution": "node",
     "allowSyntheticDefaultImports": true,
     "esModuleInterop": true,
